Fix missing quotes
Fixes a bug where the quotes would not appear after a while. The problem
was the setting of self.base_fields on the ModelForm. self.base_fields
is shared by all instances of the Form, so it leaks to other requests.
So the choices were only displayed correctly once per django process.

This also sets empty_label to None instead of removing the first choice,
which is unreliable.

Setting required=True shows a correct form validation error when the 
user submits the form without selecting anything (previously it threw a
500 error).